Major code refactor with components organization and import/export. NavBar refactor and number of itens on cart notification implemented
