+++ src/Contexts/storeContext.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const StoreContext = createContext();
+
+export { StoreContext };
